Small refactoring of built-in Matplotlib viewers so that as much as possible can be re-used for non-Qt viewers
